refactoring analyseText -> findChannels / findLinks
