+++ lib/findLinks.js
@@ -1,9 +1,7 @@
 const punctuations = ['\'', '\"', '.', ',', '!', '?', '¿', '<', '>', ':'];
-
-// scheme -> https://tools.ietf.org/html/rfc3986#section-3.1
 const schemePattern = /[a-z][a-z0-9+-.]*$/i;
 
-function analyseText(text, channelPrefixes = ['#']) {
+function findLinks(text) {
     const result = [];
     let position = 0;
     let start, end;
@@ -62,46 +60,11 @@ function analyseText(text, channelPrefixes = ['#']) {
                 });
                 position = end;
             }
-        } else if (channelPrefixes.indexOf(text[position]) !== -1 &&
-            (position === 0 || /[^0-9a-z]/i.test(text[position - 1]))) {
-            start = position;
-            end = position + 1;
-            while (end < text.length && /[^ \u0007]/.test(text[end])) {
-                end += 1;
-            }
-
-            if (end - start > 1) {
-                result.push({
-                    start: start,
-                    end: end,
-                    channel: text.slice(start, end)
-                });
-                position = end;
-            }
         }
         position += 1;
-    }
-
-	// fill
-    position = 0;
-    for (let i = 0; i < result.length; i++) {
-        const textSegment = result[i];
-        if (textSegment.start > position) {
-            result.splice(i, 0, {
-                start: position,
-                end: textSegment.start
-            });
-        }
-        position = textSegment.end;
-    }
-    if (position < text.length) {
-        result.push({
-            start: position,
-            end: text.length
-        });
     }
 
     return result;
 }
 
-module.exports = analyseText;
+module.exports = findLinks;
