+++ lib/parse.js
@@ -1,12 +1,8 @@
+const anyIntersection = require('./anyIntersection');
+const fill = require('./fill');
+const findChannels = require('./findChannels');
+const findLinks = require('./findLinks');
 const parseStyle = require('./parseStyle');
-const analyseText = require('./analyseText');
-
-function anyIntersection(a, b) {
-    return a.start <= b.start && b.start <= a.end ||
-        a.start <= b.end && b.end <= a.end ||
-        b.start <= a.start && a.start <= b.end ||
-        b.start <= a.end && a.end <= b.end;
-}
 
 function parse(text, channelPrefixes) {
     const fragments = parseStyle(text).map((fragment, i, array) => {
@@ -17,7 +13,12 @@ function parse(text, channelPrefixes) {
 
     const cleanText = fragments.map(fragment => fragment.text).join('');
 
-    const textParts = analyseText(cleanText, channelPrefixes);
+    const channelParts = findChannels(cleanText, channelPrefixes);
+    const linkParts = findLinks(cleanText);
+
+    const parts = channelParts.concat(linkParts);
+    const textParts = parts.concat(fill(parts, cleanText))
+        .sort((a, b) => a.start - b.start);
 
     return textParts.map(textPart => {
         textPart.fragments = fragments
