+++ package.json
@@ -1,20 +1,23 @@
 {
   "name": "ircmessageparser",
-  "version": "1.1.2",
+  "version": "2.0.0",
   "description": "parses irc messages and extracts styling informations, urls and channels",
   "main": "index.js",
   "directories": {
     "test": "test"
   },
-  "dependencies": {},
+  "dependencies": {
+    "escape-string-regexp": "^1.0.5"
+  },
   "devDependencies": {
     "chai": "^3.5.0",
-    "eslint": "^3.8.1",
+    "eslint": "^3.12.2",
     "istanbul": "^0.4.5",
-    "mocha": "^3.1.2",
-    "npm-run-all": "^3.1.1"
+    "mocha": "^3.2.0",
+    "npm-run-all": "^3.1.2"
   },
   "scripts": {
+    "mocha": "mocha",
     "test": "npm-run-all test:*",
     "test:lint": "eslint .",
     "test:coverage": "istanbul cover node_modules/mocha/bin/_mocha"
